@@ -1,17 +1,13 @@
 package in.bushansirgur.springboot.crudapi;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
 public class LocadoraVeiculosMainTests {
 
-	@Test
-	public void contextLoads() {
-	}
-
+    @Test
+    public void contextLoads() {
+        // Verifica se o contexto da aplicação carrega corretamente
+    }
 }
-
