@@ -1,13 +1,14 @@
 package in.bushansirgur.springboot.crudapi.dao;
 
 import java.util.List;
-import javax.persistence.EntityManager;
+
 import org.hibernate.Session;
 import org.hibernate.query.Query;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import in.bushansirgur.springboot.crudapi.model.Carro;
+import jakarta.persistence.EntityManager;
 
 @Repository
 public class CarroDAOImpl implements CarroDAO {
