@@ -2,8 +2,6 @@ package in.bushansirgur.springboot.crudapi.dao;
 
 import java.util.List;
 
-import javax.persistence.EntityManager;
-
 import org.hibernate.Session;
 import org.hibernate.query.Query;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +9,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import in.bushansirgur.springboot.crudapi.model.Funcionario;
+import jakarta.persistence.EntityManager;
 
 @Repository
 public class FuncionarioDAOimpl implements FuncionarioDAO{
